Move formatDate to helperFunction file so it can be used in App component as well
