+++ src/components/App/App.js
@@ -6,6 +6,7 @@ import { Route, Routes } from 'react-router-dom';
 import ArticleDetails from '../ArticleDetails/ArticleDetails';
 import getAllArticles from '../../apiCalls';
 import newsData from '../../newsData.json'
+import formatDate from '../../helperFunctions';
 
 
 const App = () => {
@@ -26,7 +27,7 @@ const App = () => {
 
   const updateCurrentArticle = (title) => {
     let foundArticle = articles.find(article => article.title === title)
-    const newArticle = { ...foundArticle, source: foundArticle.source.name }
+    const newArticle = { ...foundArticle, source: foundArticle.source.name, publishedAt: formatDate(foundArticle.publishedAt) }
     setCurrentArticle(newArticle)
   }
