+++ src/components/ArticleCard/ArticleCard.js
@@ -1,15 +1,8 @@
 import { Link } from 'react-router-dom'
 import './ArticleCard.css'
+import formatDate from '../../helperFunctions'
 
 const ArticleCard = ({ title, source, date, imgSrc }) => {
-  const formatDate = (data) => {
-    let month = data.slice(5, 7)
-    let day = data.slice(8, 10)
-    let year = data.slice(0, 4)
-
-    return `${month}/${day}/${year}`
-  }
-
   return (
     <Link to={`/${title}`} className='article-card' key={title} >
       <div className='card-details' >
